Modernize Tally and add budget-walk scaling experiment
Document the project structure and thesis context in SOURCE-OVERVIEW.md.

Update the GCC/CMake build, add CTest coverage, and fix current compiler/runtime issues.

Add a 10-minithread random-walk budget experiment, carry forced-yield tally debt across cycles, and generate CSV/PNG scaling plots.

+++ code/test_switch.c
@@ -16,16 +16,18 @@ int c(){
 }
 
 void t(void* args){
-    int x = (int) args;
+    int x = (int)(intptr_t) args;
     int sum = 0;
 
     switch (x)
     {
     case 1:
         sum = a();        
+        /* fall through */
 
     case 2:
         sum = b();
+        /* fall through */
 
     case 3:
         sum = c();
